Use ints instead of size_t when returning register ids

+++ inc/watchdog.h
@@ -16,9 +16,9 @@ typedef struct WatchdogPack WatchdogPack;
 WatchdogPack* watchdogpack_create(const size_t wdogs_number);
 size_t watchdogpack_get_registered(register const WatchdogPack* wdog_pack);
 size_t watchdogpack_get_size(register const WatchdogPack* wdog_pack);
-size_t watchdogpack_register(WatchdogPack* wdog_pack, register const Watchdog* wdog);
+int watchdogpack_register(WatchdogPack* wdog_pack, register const Watchdog* wdog);
 void watchdogpack_unregister(WatchdogPack* wdog_pack, size_t wdog_id);
 void watchdogpack_destroy(WatchdogPack* wdog_pack);
-size_t watchdogpack_check_alarms(const WatchdogPack* wdog_pack);
+int watchdogpack_check_alarms(const WatchdogPack* wdog_pack);
 
 #endif // !WATCHDOG_H
